+++ frontend/src/saved-trips/SavedTrip.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useReducer } from 'react';
+import React, { useState, useEffect } from 'react';
 import ListGroup from 'react-bootstrap/ListGroup';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPencilAlt } from '@fortawesome/free-solid-svg-icons';
@@ -13,33 +13,7 @@ import ShareTripModal from '../share-trip/ShareTripModal.js';
  */
 function SavedTrip({ tripId }) {
   const [showShareModal, setShowShareModal] = useState(false);
-
-  function setTripObjectDispatch(state, action) {
-    const parsedAttractions = [];
-    const attractions = JSON.parse(action.attractions);
-    attractions.forEach((attr) => {
-      const parsedAttr = {
-        name: JSON.parse(attr.name),
-        photoUrl: JSON.parse(attr.photoUrl),
-        lat: JSON.parse(attr.lat),
-        lng: JSON.parse(attr.lng),
-        routeIndex: attr.routeIndex,
-      };
-      parsedAttractions.push(parsedAttr);
-    });
-    return {
-      tripId: state.tripId,
-      attractions: parsedAttractions,
-      centerLocation: JSON.parse(action.centerLocation),
-      isOptimized: action.isOptimized,
-      searchText: JSON.parse(action.searchText),
-      tripName: JSON.parse(action.tripName),
-    };
-  }
-  const [tripObject, setTripObject] = useReducer(setTripObjectDispatch, {
-    tripId: tripId,
-    attractions: [],
-  });
+  const [tripObject, setTripObject] = useState({});
 
   useEffect(() => {
     getTrip(tripId).then((trip) => setTripObject(trip));
